% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dflow.R
\name{dflow}
\alias{dflow}
\title{Design Flow Calculation}
\usage{
dflow(
  x,
  m,
  r,
  yearstart = NA,
  yearend = NA,
  wystart = "10-01",
  wyend = "09-30"
)
}
\arguments{
\item{x}{Data frame where col 1 = POSIXct date and col 2 = numeric daily mean flow}

\item{m}{Flow averaging period in days}

\item{r}{Return period in years}

\item{yearstart}{Optional. The starting year of the calculation period.
Should be in integer format as yyyy. If not specified the
default will be the year of the minimum date in x.}

\item{yearend}{Optional. The ending year of the calculation period.
Should be in integer format as yyyy. If not specified the
default will be the year of the maximum date in x.}

\item{wystart}{Optional. Character date (excluding year) that begins the water year in format "mm-dd".
Default is "10-01"}

\item{wyend}{Optional. Character date (excluding year) that ends the water year in format "mm-dd".
Default is "09-30".}
}
\value{
numeric design flow.
}
\description{
Function to find the design flow based on methodology used in U.S. EPA's \href{https://www.epa.gov/ceam/dflow}{DFLOW program}
with a modification to account for missing flow data or days with zero flow.
If there are missing flow data in any particular water year, all data from that water year is not used.
This approach is \href{https://www.epa.gov/ceam/technical-support-dflow#xqy}{recommended by EPA} (although not implemented in DFLOW 3.1)
and applied in USGS's \href{https://water.usgs.gov/software/SWSTAT/}{SWSTAT} program.
}
\details{
The methodology for this function is based on the equations contained in DFLOW user manual
as presented by Rossman (1999). Design flow is computed from the sample of lowest m-day
average flow for each defined year, where "m" is the user supplied flow averaging period.
The arithmetic averaging is used to calculate the m-day average flows. A log Pearson Type III
probability distribution is fitted to the annual minimum m-day flows. The design
flow is the value from the distribution whose probability of not being exceeded is 1/r where r
is the user-supplied return period.

If a particular season is desired, set wystart and wyend arguments to correspond to the beginning
and ending dates of the season. The season does not have to span an entire year. Only flow records
that fall within the season will be used in the calculations.

References:

Rossman, L A. 1990. DFLOW USER'S MANUAL. U.S. Environmental Protection Agency,
Washington, DC, EPA/600/8-90/051 (NTIS 90-225616). https://nepis.epa.gov/Exe/ZyPDF.cgi/30001JEH.PDF?Dockey=30001JEH.PDF

DFLOW: https://www.epa.gov/ceam/dflow
}
